+++ sdk/core/azure-core/azure/core/pipeline/policies/_retry.py
@@ -80,7 +80,7 @@ class RetryPolicy(HTTPPolicy):
 
     :keyword RetryMode retry_mode: Fixed or exponential delay between attemps, default is exponential.
 
-    :keyword int max_timeout: Maximal timeout setting for the operation in seconds, default is 600s (10 minutes).
+    :keyword int timeout: Timeout setting for the operation in seconds, default is 600s (10 minutes).
 
     .. admonition:: Example:
 
@@ -107,7 +107,7 @@ class RetryPolicy(HTTPPolicy):
         self.backoff_factor = kwargs.pop('retry_backoff_factor', 0.8)
         self.backoff_max = kwargs.pop('retry_backoff_max', self.BACKOFF_MAX)
         self.retry_mode = kwargs.pop('retry_mode', RetryMode.Exponential)
-        self.max_timeout = kwargs.pop('max_timeout', 600)
+        self.timeout = kwargs.pop('timeout', 600)
 
         retry_codes = self._RETRY_CODES
         status_codes = kwargs.pop('retry_on_status_codes', [])
@@ -137,7 +137,7 @@ class RetryPolicy(HTTPPolicy):
             'backoff': options.pop("retry_backoff_factor", self.backoff_factor),
             'max_backoff': options.pop("retry_backoff_max", self.BACKOFF_MAX),
             'methods': options.pop("retry_on_methods", self._method_whitelist),
-            'max_timeout': options.pop("max_timeout", self.max_timeout),
+            'timeout': options.pop("timeout", self.timeout),
             'history': []
         }
 
@@ -299,12 +299,6 @@ class RetryPolicy(HTTPPolicy):
         :return: False if have more retries. True if retries exhausted.
         :rtype: bool
         """
-        max_timeout = settings['max_timeout']
-        start_time = settings['start_time']
-        time_consuption = time.time() - start_time
-        if time_consuption >= max_timeout:
-            return True
-
         retry_counts = (settings['total'], settings['connect'], settings['read'], settings['status'])
         retry_counts = list(filter(None, retry_counts))
         if not retry_counts:
@@ -389,7 +383,7 @@ class RetryPolicy(HTTPPolicy):
         if retry_settings['history']:
             context['history'] = retry_settings['history']
 
-    def send(self, request):
+    def send(self, request):    # pylint: disable=too-many-statements
         """Sends the PipelineRequest object to the next policy. Uses retry settings if necessary.
 
         :param request: The PipelineRequest object
@@ -424,10 +418,15 @@ class RetryPolicy(HTTPPolicy):
 
         retry_settings['body_position'] = body_position
         retry_settings['file_positions'] = file_positions
-        retry_settings['start_time'] = time.time()
+
+        absolute_timeout = retry_settings['timeout']
 
         while retry_active:
             try:
+                start_time = time.time()
+                if absolute_timeout <= 0:
+                    break
+                request.context.options['connection_timeout'] = absolute_timeout
                 response = self.next.send(request)
                 if self.is_retry(retry_settings, response):
                     retry_active = self.increment(retry_settings, response=response)
@@ -446,6 +445,10 @@ class RetryPolicy(HTTPPolicy):
                         self.sleep(retry_settings, request.context.transport)
                         continue
                 raise err
+            finally:
+                end_time = time.time()
+                if absolute_timeout:
+                    absolute_timeout -= (end_time - start_time)
 
         self.update_context(response.context, retry_settings)
         return response